packages/ai-server/pyproject.toml
[project]
name = "decoded-ai"
version = "1.0.0"
description = ""
authors = [
    {name = "CIOI",email = "rhkr9693@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "openai (>=1.65.2,<2.0.0)",
    "langdetect (>=1.0.9,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "dependency-injector (>=4.46.0,<5.0.0)",
    "bson (>=0.5.10,<0.6.0)",
    "selenium (>=4.29.0,<5.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "fake-useragent (>=2.0.3,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "toolz (>=1.0.0,<2.0.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "grpcio>=1.74.0",
    "grpcio-tools>=1.74.0",
    "protobuf>=6.31.1",
    "torch (>=2.7.1,<3.0.0)",
    "transformers (>=4.53.2,<5.0.0)",
    "python-telegram-bot (>=21.0.0,<22.0.0)",
    "youtube-transcript-api (>=1.2.0,<2.0.0)",
    "groq (>=0.31.0,<0.40)",
    "jinja2 (>=3.1.0,<4.0.0)",
    "google-genai (>=1.10.0,<1.55.0)",  # <1.55: extra_items(TypedDict)는 Python 3.15에서 추가
    "apscheduler (>=3.10.0,<4.0.0)",
    "arq (>=0.26.0,<0.27.0)",
    "perplexityai (>=0.27.0,<0.28.0)",
    "langgraph (>=0.3)",
    "langchain-groq (>=0.2)",
    "langchain-core (>=0.3)",
    "supabase (>=2.0)",
    "tenacity (>=9.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
ipykernel = "^6.29.5"
poethepoet = "^0.33.0"
pillow = "^11.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
test = "pytest"
lint = "flake8"
format = "black ."